mới nhất
chưa sửa xong lỗi giá

+++ da_be/src/main/java/com/example/da_be/response/HoaDonKHResponse.java
@@ -18,7 +18,7 @@ public class HoaDonKHResponse {
     private BigDecimal phiShip;
     private String sanPhamTen;
     private BigDecimal giaBan;
-    private Integer giaKhuyenMai;
+    private BigDecimal giaKhuyenMai;
     private Integer soLuongMua;
     private String hinhAnhLink;
     private Integer giaTriVoucher;
